+++ praw/__init__.py
@@ -45,9 +45,18 @@ from warnings import warn_explicit
 
 
 __version__ = '2.1.14'
+
+if os.environ.get('SERVER_SOFTWARE') is not None:
+    # Google App Engine information
+    # https://developers.google.com/appengine/docs/python/
+    platform_info = os.environ.get('SERVER_SOFTWARE')
+else:
+    # Standard platform information
+    platform_info = platform.platform(True)
+
 UA_STRING = '%%s PRAW/%s Python/%s %s' % (__version__,
                                           sys.version.split()[0],
-                                          platform.platform(True))
+                                          platform_info)
 
 MIN_IMAGE_SIZE = 128
 MAX_IMAGE_SIZE = 512000